windows_monitor/requirements.txt
psutil>=5.8.0
pyserial>=3.5
GPUtil>=1.4.0